+++ src/main/java/net/frozenblock/lib/levelgen/structure/impl/status/StructureStatusUpdater.java
@@ -17,14 +17,12 @@
 
 package net.frozenblock.lib.levelgen.structure.impl.status;
 
-import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
-import net.frozenblock.lib.FrozenLibConstants;
-import net.frozenblock.lib.levelgen.structure.impl.StructureStartInterface;
+import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Holder;
-import net.minecraft.resources.Identifier;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.StructureManager;
@@ -36,36 +34,26 @@ import org.jetbrains.annotations.ApiStatus;
 public class StructureStatusUpdater {
 
 	public static void updatePlayerStructureStatusesForLevel(ServerLevel level) {
-		final StructureManager structureManager = level.structureManager();
-		level.players().forEach(player -> updatePlayerStructureStatus(structureManager, player));
+		PlayerLookup.level(level).forEach(player -> {
+			if (!player.connection.hasClientLoaded()) return;
+			updatePlayerStructureStatus(level.registryAccess().lookupOrThrow(Registries.STRUCTURE), level.structureManager(), player);
+		});
 	}
 
-	private static void updatePlayerStructureStatus(StructureManager structureManager, ServerPlayer player) {
+	private static void updatePlayerStructureStatus(Registry<Structure> structureRegistry, StructureManager structureManager, ServerPlayer player) {
 		final BlockPos pos = player.blockPosition();
 		final List<StructureStatus> newStructureStatuses = new ArrayList<>();
-		final List<StructureStatus> currentStructureStatuses = player.getAttachedOrElse(StructureStatus.ATTACHMENT_TYPE, ImmutableList.of());
 
 		for (Structure structure : structureManager.getAllStructuresAt(pos).keySet()) {
 			final StructureStart structureStart = structureManager.getStructureAt(pos, structure);
 			if (structureStart == StructureStart.INVALID_START) continue;
 
-			if (!((Object) structureStart instanceof StructureStartInterface structureStartInterface)) continue;
-
-			final Identifier structureLocation = structureStartInterface.frozenLib$getId();
-			if (structureLocation != null) {
-				boolean insidePiece = structureManager.structureHasPieceAt(pos, structureStart);
-				boolean addNewStructureStatus = true;
-				for (StructureStatus existingStatus : newStructureStatuses) {
-					if (!existingStatus.structure().is(structureLocation)) continue;
-					addNewStructureStatus = existingStatus.insidePiece() != insidePiece;
-				}
-				if (addNewStructureStatus) newStructureStatuses.add(new StructureStatus(Holder.direct(structure), insidePiece));
-			} else if (FrozenLibConstants.UNSTABLE_LOGGING) {
-				throw new AssertionError("Structure piece doesn't contain an id!");
-			}
+			final boolean insidePiece = structureManager.structureHasPieceAt(pos, structureStart);
+			newStructureStatuses.add(new StructureStatus(structureRegistry.getKey(structure), insidePiece));
 		}
 
-		if (newStructureStatuses.equals(currentStructureStatuses)) return;
-		player.setAttached(StructureStatus.ATTACHMENT_TYPE, newStructureStatuses);
+		if (!newStructureStatuses.equals(player.getAttached(StructureStatus.ATTACHMENT_TYPE))) {
+			player.setAttached(StructureStatus.ATTACHMENT_TYPE, newStructureStatuses);
+		}
 	}
 }